% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_summary_table.R
\name{describe_summary_table}
\alias{describe_summary_table}
\title{Column dictionary for build_summary_table() output}
\usage{
describe_summary_table()
}
\description{
Column dictionary for build_summary_table() output
}
