% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_to_edgeR.R
\name{phyloseq_to_edgeR}
\alias{phyloseq_to_edgeR}
\title{Convert a phyloseq object to an edgeR DGEList}
\usage{
phyloseq_to_edgeR(physeq, group, method = "RLE", ...)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{group}{Character (a sample_data column name) or a vector of group
labels, one per sample.}

\item{method}{Normalization method passed to edgeR::calcNormFactors().}

\item{...}{Passed to edgeR::DGEList().}
}
\value{
An edgeR DGEList with dispersions estimated.
}
\description{
Convert a phyloseq object to an edgeR DGEList
}
