% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTUs_multi_DA.R
\name{consolidate_DA_results}
\alias{consolidate_DA_results}
\title{Consolidate Differential Abundance Results}
\usage{
consolidate_DA_results(
  edgeR_res = NULL,
  DESeq2_res = NULL,
  ALDEx2_res = NULL,
  metagenomeSeq_res = NULL,
  ADAPT_res = NULL
)
}
\arguments{
\item{edgeR_res}{EdgeR results}

\item{DESeq2_res}{DESeq2 results}

\item{ALDEx2_res}{ALDEx2 results}

\item{metagenomeSeq_res}{metagenomeSeq results}

\item{ADAPT_res}{ADAPT results}
}
\description{
Consolidate Differential Abundance Results
}
